+++ gatling-charts-highcharts/src/main/scala/com/excilys/ebi/gatling/highcharts/series/ScatterSeries.scala
@@ -8,7 +8,7 @@ package com.excilys.ebi.gatling.highcharts.series
 import com.excilys.ebi.gatling.core.util.StringHelper.END_OF_LINE
 import scala.collection.mutable.ArrayBuffer
 
-class ScatterSeries(name: String, data: List[(Int, Int)], color: String) extends HighchartsSeries[Int, Int](name, data, List(color)) {
+class ScatterSeries(name: String, data: List[(Int, Long)], color: String) extends HighchartsSeries[Int, Long](name, data, List(color)) {
 
 	def getElements: ArrayBuffer[String] = {
 		val buffer = new ArrayBuffer[String]
